Add agent into simulation

+++ Assets/src/model/indoor_sim/AgentDescriptor.cs
@@ -8,9 +8,9 @@ public class AgentDescriptor
 {
     public string name = "";
     public string type = "";
-    public double x;
-    public double y;
-    public double theta;
+    public float x;
+    public float y;
+    public float theta;
     public string? containerId;  // if containerId != null, them there should be one container with this id in IndoorData.
                                  // And the x, y will be the Centroid of the container.geom
